Refactor how cache is managed within the application
Create an interface for the CacheManager, with an implementation that
uses Gemfire (along with a stub for testing).

The CacheManager is passed into the HTTPServer instead of the server
creating an instance of a cache manager.

HTTPServer no longer needs to know anything about the inner workings of
the cache (eg. region). If it needs something done in relation to the
cache, it gets the cache manager to handle it.

At this point, the server still does not use the cache for handling user
requests and still goes to the 'slow' backend repository.